The useful links and the recycling symbols options in the drawer now work. Splash screen has been altered slightly (#8)
In the recycling symbols option we have a horizontal scroll view, when u touch an image there is a vibrate and a toast comes up to tell you what the symbol means. I have put four links on the useful links option , when you click the link it takes you to the webpage. Made a very small alteration on the splashscreen as it was staying open for far too long.