Claudine (#11)
* locations

locations class

* map activity (again)

map map maps

* Map Activity

So it worked on one emulator, but itsnt working on my phone when I run the app, but just commiting what I have so far and then will try to fix it.

* Map working

Okay so it finally works when you click show all recycling centres... only we dont have a way of showing them yet! But work in progress :+1:

* More Map stuff

I somehow set the background of main activity to be a map but have fixed that with this commit. Also added a search bar to the map.. which isnt fully funtional just yet but almost!