Bin companies (#19)
* Bin companies option

ok, so i've coded up the bin companies option in the drawer. It jumps to a new activity where we have the name, number and a link to their website. I have also added a customised button. When this is clicked it takes you to the phone where the number is displayed and you can decide whether you want to call or not. I decided to do it that way, rather than the  automatically call cos sometimes that can catch you out!

* back buttons

back buttons added to the action bar for all activities that needed it

* back buttons

back buttons added to the action bar for all activities that needed it