Claudine (#28)
* locations

locations class

* map activity (again)

map map maps

* Map Activity

So it worked on one emulator, but itsnt working on my phone when I run the app, but just commiting what I have so far and then will try to fix it.

* Map working

Okay so it finally works when you click show all recycling centres... only we dont have a way of showing them yet! But work in progress :+1:

* More Map stuff

I somehow set the background of main activity to be a map but have fixed that with this commit. Also added a search bar to the map.. which isnt fully funtional just yet but almost!

* So Ive added markers which appear after a location is searched. Also added a button which recenters the map on your location. The search function should work with no problems. Next step is to add Google Places Api which will allow auto suggestions to appear and give more details when the marker is clicked. (Note: this works on my phone but crashes on the emulators for some reason)

* about

Added about page

* Updated About

added a background image and played around with some text.

* minor changes

Got rid of the map toasting message. Also fixed the white bar on the about screen.

* adds back button to map

adds the code from lorraine's child activity giving a back button to the map.

* fixed map camera zoom on starting location

also added a few comments